[testDBClientHatohol][testDBTermCodec] Fix crashed tests on 32bit env

+++ server/test/testDBTermCodec.cc
@@ -61,19 +61,19 @@ void test_getInt(gconstpointer data)
 void data_getUint64(void)
 {
 	gcut_add_datum("Zero",
-		       "val", G_TYPE_UINT64, 0,
+		       "val", G_TYPE_UINT64, (guint64)0,
 		       "expect", G_TYPE_STRING, "0",
 		       NULL);
 	gcut_add_datum("Positive within 32bit",
-		       "val", G_TYPE_UINT64, 3456,
+		       "val", G_TYPE_UINT64, (guint64)3456,
 		       "expect", G_TYPE_STRING, "3456",
 		       NULL);
 	gcut_add_datum("Positive 32bit Max",
-		       "val", G_TYPE_UINT64, 2147483647,
+		       "val", G_TYPE_UINT64, (guint64)2147483647,
 		       "expect", G_TYPE_STRING, "2147483647",
 		       NULL);
 	gcut_add_datum("Positive 32bit Max + 1",
-		       "val", G_TYPE_UINT64, 2147483648,
+		       "val", G_TYPE_UINT64, (guint64)2147483648,
 		       "expect", G_TYPE_STRING, "2147483648",
 		       NULL);
 	gcut_add_datum("Positive 64bit Poistive Max",
@@ -105,19 +105,19 @@ namespace testDBTermCodecSQLite3 {
 void data_getUint64(void)
 {
 	gcut_add_datum("Zero",
-		       "val", G_TYPE_UINT64, 0,
+		       "val", G_TYPE_UINT64, (guint64)0,
 		       "expect", G_TYPE_STRING, "0",
 		       NULL);
 	gcut_add_datum("Positive within 32bit",
-		       "val", G_TYPE_UINT64, 3456,
+		       "val", G_TYPE_UINT64, (guint64)3456,
 		       "expect", G_TYPE_STRING, "3456",
 		       NULL);
 	gcut_add_datum("Positive 32bit Max",
-		       "val", G_TYPE_UINT64, 2147483647,
+		       "val", G_TYPE_UINT64, (guint64)2147483647,
 		       "expect", G_TYPE_STRING, "2147483647",
 		       NULL);
 	gcut_add_datum("Positive 32bit Max + 1",
-		       "val", G_TYPE_UINT64, 2147483648,
+		       "val", G_TYPE_UINT64, (guint64)2147483648,
 		       "expect", G_TYPE_STRING, "2147483648",
 		       NULL);
 	gcut_add_datum("Positive 64bit Poistive Max",
